Update the project's `readme.md`
This commit contains the `markdown` source code for the enhanced
`readme.md`. This commit adds the missing sections that have been
defined in the issue: #3, and similarly mentioned in #18.

Moreover, I've left a set of `TODO` messages that signify a further
improvement in a place that they have been written in, such as the
`Installation`, `Usage` sections, etc.