Parse loudness value via `mapToPercentage()` function
Previously, we had a function `calculatePercentage()` which did the same
as the `map()` function, albeit returning a `float` type value. However,
we've realized that the difference between having an `int` and a `float`
is not too significant, so we make reuse of the `mapToPercentage()`
function instead (for the loudness calculation).