#26 Fix incorrect mapping for moisture sensor
The mapping's max. value was incorrect. This commit fixes the max.
value. As a consequence, the `enum` for the types of sensors is omitted
(it's not required in any functionality now). The `mapToPercentage()`
function does not, consequently, require the 2nd parameter with the
sensor type, thus reducing the function's body to just a return
statement via the `map()` function (the `switch` statement is fully
omittable). Similarly, the macro with the incorrect max. value is
removed.