#24 Display UI in Wio LCD Screen and enable Buttons
This commit creates the UI for displaying each individual sensor data as well as the dashboard containing all the data in parallel. It also enables the use of the Wio Terminal's buttons for switching between these screens.

It does not contain integration with actual sensor data readings and the code contained in terminarium.ino. This is just a test file that works with static test values that change only by simulation.

Future commits will refactor the code greatly (currently primitive, with much repetition), add thorough comments, and integrate with our main terminarium.ino file.