#27 Add `Shell` script to emulate the system in CLI
This commit adds a shell script (suitable for Unix-based) system, which uploads the sketch to the Wio Terminal, as well as reads the data from the serial monitor onto a terminal emulator. Moreover, the script can prompt the user with the intended usage, as well as list the available ports on the machine. The main goal of the script is to enable users to interact with the Terminarium system from the command line.