#21 Add initial `CONTRIBUTING.md`
This commit adds the initial version of the `CONTRIBUTING.md` document. It is used to ensure that team members are following the same conventions in terms of: creating issues, writing commits, documenting/accepting/handling merge requests, using appropriate GitLab labels, and providing milestones and/or due dates to requirements. The document is written in the `Markdown` syntax.