#29 Add `Vue.js` initial configuration and default **project scaffolding**
This commit adds the proper initial configuration of a Vuejs project (with Vite) with the required npm dependencies. Similarly, it adds the default Vue3 project scaffolding. Note: the scaffolding is subject to change, or to be removed completely. Still, as of now, it is preserved, to ensure that all team members are able to operate in the Vue environment on their machines, hence avoiding lots of potential issues in the future.