#31 Add initial `node` env. config
This commit adds the initial configuration for the node modules. The modules are used to define back-end utilities that will be later coupled with the UI portion of the website. Currently, the dotenv, firebase packages are defined as the dependencies. Indeed, the scope of the dependencies is subject to evolve.