#30 Establish MQTT connection between Wio Terminal and computer
This commit establishes mqtt connection between the wio terminal and computer so that the web app can receive sensor data.

It also changes the code for the LCD display from using screen buffer sprite to writing directly on the screen, because the wifi and pubsubclient libraries are very memory intensive and could not be used with a screen buffer.

This commit also makes some minor readability and formatting changes.

Known bugs include LCD screen flickering every time it refreshes.