#31 Connect to database
This commit adds the initial version of the `database.js` module that communicates with the chosen database (i.e. Firebase Realtime Database). It uses the configuration defined at a secret .env file and attempts to connect to the DB. If a connection is successful, a listener function `onValue` will detect that, and report it to the console by printing a String literal to signify the status update.