#61 Read sensor data given a time range
This commit adds a stand-alone JavaScript module that connects to the Database instance and fetches (i.e. reads) entries within a defined time range. That is, if hours = 10, all the entries that were registered within the time range of {(now - 10h), now} are to be read. As of now, upon reading, the result entries are, simply, printed (in a formatted manner) to the console. Later, this module will be part of the website-application and the results are to be dynamically displayed in the corresponding components.