#43 Add 'DefaultPageLayout'-component
This commit adds the 'DefaultPageLayout'-component that contains the basic elements (also individual components) which will always be displayed on the website - it also refactors the 'App.vue'-file to make use of the newly created component. By breaking the website up into modular components, we can easily make changes in one place which will then be reflected wherever the component is used, as well as let us dynamically choose what to display to the end user.