#41 Complete connectivity screen functionality
This commit introduces new screens allowing user to opt in to mqtt connection

This includes a connection selection screen, a wifi connection screen and a mqtt connection screen, and display text that changes depending on context.

It also includes icons in the header that indicate whether wifi and mqtt are connected.

Numerous new functions have been added in service of the above, including draw functions and screen control logic.