#65 Migrate fetch.js module to website's source; make changes
This commit migrates the fetch.js module (used for reading from the database) under src/modules/ of the website. Similarly, the module has been updated in a way, such that in can be used within the Vue environment (via Vite). Furthermore, now, the function therein results Map(sensor: {...}). Previously, it was Array({...}). This way, it's easier to access the specific values given a sensor name.