#64 Add Vue component for Graph(s)
This commit adds the source of the Vue component for the Graph(s) namely via Charts.vue under the components folder. The component enhances reusability; moreover, it has a mechanism that can automatically assign a new Graph instance a unique ID. Furthemore, the Chart component accepts a range value which is then used to traverse the DB within a time interval (in order to obtain data entries). Ultimately, the dimensions of the Graph can be specified too.