#64 Fix props. for sensorType, add documentation
This commit changes the props. Object for the Chart Component, namely for the sensorType field. Previously, by a mistake, the field was `default: true` which was replaced by the correct notation: `required: true` (to indicate that the field is required to be given). Moreover, there's an additional comment adduced that lists the expected values for the sensorType.