#81 Add testing job to GitLab pipeline
This commit adds the test-job to the pipeline. It relies on the 'npm run test:unit' command of the npm environment. It performs the tests via Unit Testing on the Vue components. The test-job takes place the 3rd in the order; before building, after linting. The job is added to the '.gitlab-ci.yml' file from the root of the repository.