#70 Write vibrationCount to database and reset each interval
This commit modifies the saveData function in dataHandler to increment the vibration counter by 1 every time it receives a message from the vibration sensor equating to 'true'.

It also modifies the writeToDB function so that this vibrationCount is written each interval (and then reset) rather than the real-time true or false data from the sensorValues map.