#85 Add initial Unit Test for AccessButton
This commit adds the initial source for the Unit Test of the AccessButton component. It tests the core functionalities of the component, namely rendering, structure, and logic. However, since the component lacks the implementation of the function that redirects the user based on the 'log in state', there's a 'TODO' message left to signify that such a test is subject to be added.