#90 Add confirm network credentials screen and modify buttons
This commit adds a screen that appears after choosing to establish mqtt connection.

The screen allows user to confirm network details, and modify if necessary.

It also transfers the functionalitiy of the top 3 buttons of the Wio Terminal to the joystick at the bottom due to technical issues involving using both sets at once.