#91 New fix using millis()
This commit modifies the previous fix with a new one that allows the update screen to behave correctly regardless of how many mqtt messages are received, both isolated or in succession.

It also restructures the main Terminarium.ino so that the mqtt listening function is placed outside the code block that runs only 1nce per interval.

This is done to increase responsiveness when updating sensor ranges.

Lastly the goPrevScreen() function is removed as it is obsolete.