#71 Add Monitoring Component to 'SelectTerrarium' page
This commit integrates the 'Chart'-component into the 'MonitorTerrarium'-component to display the graphs generated from the database data. It also re-factors the file, adding structure and code comments for readability. Furthermore, this commit also modifies the SelectTerrariumPage-component, since due to a change in scope/design, the terrarium monitoring is now part of the former, as opposed to making up its own unique page on the website. Co-authored-by: James Klouda @klouda