#72 Update Header Unit test with routing testing suite
This commit updates the Unit Test for the Header component. Now, it performs a check which ensures that the 'routing' is in place and in the desired format. Moreover, the test is defined in a way that the isLoggedIn prop. is mocked as if it was injected from the parent. Co-authored-by: Erik Lindstrand elindstr@student.chalmers.se