#97 Update AccessButton Unit test
This commit updates the Unit Test for the AccessButton component due to a recent changes in the structure of the components, namely due to the finalisation of the 'routing' based on the 'hashchange' principle. Moreover, the TODO message is removed, since the 'redirecting' (upon clicking on the button) is not handled by the component based on a recent agreement.