#71 Refactor Monitoring-related Components
This commit contains several changes to the affected components: adds the 'Set Ranges'-button to the 'MonitorTerrarium'-component, which allows the user to set the acceptable sensor ranges; alters the styling of the elements displayed on the 'SelectTerrarium'-screen to improve readability; fixes missing forward slashes in image source path-notations within the 'SelectTerrariumPage' and 'Avatar' components which would lead to image assets not being displayed properly in the live deployed version of the website. Additionally, more code comments have been added to improve readability/understanding for other developers of this project.